Convert some symbolic floats to rationals on simplification.
Monotone-Parent: 3a79069708e3b8296f84ec6e78c6237a63f6e584
Monotone-Revision: 3e7dbd00d6f5e0fc761fa64e495b3b909b51f15b

+++ include/simple_cfd/symbolic/rational.hpp
@@ -20,15 +20,15 @@ class Rational : public AbstractBasic<Rational>,
                  boost::equality_comparable<Rational>
 {
 private:
-  int numerator;
-  int denominator;
+  long numerator;
+  long denominator;
 
   void normalise();
-  static unsigned gcd(unsigned a, unsigned b);
+  static unsigned long gcd(unsigned long a, unsigned long b);
 
 public:
-  Rational(int value);
-  Rational(int numerator, int denominator);
+  Rational(long value);
+  Rational(long numerator, long denominator);
   virtual std::size_t nops() const;
   virtual void write(std::ostream& o) const;
   virtual Expr derivative(const Symbol& s) const;
